vhost-scsi: add new rpc to enhance live recovery
After the vhost-scsi controller is started, qemu can reconnect
to spdk, but at this time the LUN may not be added to the scsi
target, struct spdk_scsi_dev_vhost_state's member dev has not
been set yet, which will lead to IO error subsequently.

So we add a new rpc to defer starting the vhost-scsi controller
until adding the scsi target is completed.

Change-Id: I1f8a78528de3073293d1c5ce2c064123519bc320
Signed-off-by: Haoqian He <haoqian.he@smartx.com>