lib: add spdk_mem_(un)register_ext for memory register
Currently spdk_mem_register registers memory will
executes hook handlers unconditionally.
For example, if both rdma and vhost-user are enabled,
the VM memory will also be registered to rdma mr.
However if there are many VMs with large RAM, the rdma
mr register will be too heavy, which increasing the
time of the vhost-user connection, even if the rdma
mr register operation is not necessary.

In fact, not all memory needs to execute vfio-user and
rdma hook operations. So this patch provides an interface
for more precise memory register control.

Signed-off-by: Haoqian He <haoqian.he@smartx.com>